New class CombinedChart for map with interpolation.
